wayland: Traverse the GdkWindow tree to find real parent
The rectangle in GdkAttachParams is relative to the GdkWindow in
GdkAttachParams. That window may be a real window, but a virtual window
used internally by GDK and GTK+. Lets traverse the tree, and add the
relative positions until we found the native window, and append that to
the attach rectangle.